refactor(UserChallenges): renaming form, updating routing to UserChallenges
