+++ src/modules/HomeNavigation/HomeNavigation.tsx
@@ -86,7 +86,7 @@ const HomeNavigation = () => {
       <Stack.Screen name={HomeNavigationRoutes.Education} component={Education} />
       <Stack.Screen name={HomeNavigationRoutes.Skills} component={Skills} />
       <Stack.Screen name={HomeNavigationRoutes.MyCourses} component={MyCourses} />
-      <Stack.Screen name={HomeNavigationRoutes.MyChallenges} component={UserChallenges} />
+      <Stack.Screen name={HomeNavigationRoutes.UserChallenges} component={UserChallenges} />
     </Stack.Navigator>
   )
 }
